requirements.txt
discord.py>=1.7.3
python-dotenv>=0.15.0
pytest>=6.0.0
pytest-asyncio>=0.15.1
